move websocket code out of base_server
