+++ dash/backends/_quart.py
@@ -45,6 +45,8 @@ from .base_server import (
     BaseDashServer,
     RequestAdapter,
     ResponseAdapter,
+)
+from .ws import (
     DashWebsocketCallback,
     run_ws_sender,
     run_callback_in_executor,
